+++ lib/ui/pages/home/index.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 import 'package:first_app/models/appstate.dart';
 import 'package:first_app/models/locstate.dart';
@@ -6,6 +7,7 @@ import 'package:first_app/ui/pages/login/index.dart';
 import 'package:first_app/generated/l10n.dart';
 import 'package:first_app/ui/pages/location/index.dart';
 import 'package:first_app/ui/pages/map/index.dart';
+import 'drawer.dart';
 
 class HomePage extends StatelessWidget {
   @override
@@ -16,34 +18,31 @@ class HomePage extends StatelessWidget {
         body: LoginPage(),
       );
     }
+
+    // as documented in appstate.dart, we here set the defaultLokale
+    // from appState to apply loaded locale from sharedpreferences on
+    // startup.
+    Intl.defaultLocale = appState.locale;
+
     return Scaffold(
-      appBar: AppBar(
-        title: Text(S.of(context).appTitle),
-      ),
-      backgroundColor: Theme.of(context).backgroundColor,
-      body: new ChangeNotifierProvider(
-        create: (_) => new LocStateModel(appState.mocks.getMock('geolocator')),
-        child: Center(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              LocationStreamWidget(),
-              OverlayMapPage(),
-            ],
+        key: Key("HomePage_Scaffold"),
+        appBar: AppBar(
+          title: Text(S.of(context).appTitle),
+        ),
+        backgroundColor: Theme.of(context).backgroundColor,
+        body: new ChangeNotifierProvider(
+          create: (_) =>
+              new LocStateModel(appState.mocks.getMock('geolocator')),
+          child: Center(
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                LocationStreamWidget(),
+                OverlayMapPage(),
+              ],
+            ),
           ),
         ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: appState.logOut,
-        tooltip: S.of(context).logoutButton,
-        key: Key('HomePage_ExitButton'),
-        child: Icon(
-          Icons.exit_to_app,
-          color: Color(0xe81751ff),
-        ),
-        backgroundColor: Theme.of(context).buttonColor,
-        foregroundColor: Theme.of(context).focusColor,
-      ),
-    );
+        drawer: HomePageDrawer());
   }
 }